app/requirements.txt
fastapi
pydantic
requests
python-dotenv
httpx
